+++ CHANGELOG.md
@@ -1,11 +1,19 @@
 # Changelog
 
+## 1.05.0
+**НАЧИНАЯ С ЭТОЙ ВЕРСИИ ХРАНЯТСЯ И ИСПОЛЬЗУЮТСЯ ТОЛЬКО ИЗОБРАЖЕНИЯ ОРИГИНАЛЬНОГО РАЗМЕРА!!!**
+Если для Вас важно сохранить старые изображения и вы использовали версии ДО 1.03.0, 
+то перенесите недостающие файлы из каталога scalable в original
+
+- Fit and pad image mode.  [#19](https://github.com/maxifly/ImageServer/issues/19
+- Save only original image size. [#17](https://github.com/maxifly/ImageServer/issues/17
+
 ## 1.04.0
 
 - Save image in original size.  [#9](https://github.com/maxifly/ImageServer/issues/9)
 - Add local image provider. [#13](https://github.com/maxifly/ImageServer/issues/13), [#15](https://github.com/maxifly/ImageServer/issues/15)
 - Fix clear directory. [#18](https://github.com/maxifly/ImageServer/issues/18)
-- 
+
 ## 1.03.0
   **ТРЕБУЕТСЯ ПРОШИТЬ РАМКУ**: версией 1.03
 - Use https.  [#4](https://github.com/maxifly/ImageServer/issues/4)